+++ common/src/main/java/net/mehvahdjukaar/polytone/utils/LegacyHelper.java
@@ -507,7 +507,7 @@ public class LegacyHelper {
             }
             if (fogCol != null || skyCol != null) {
                 var mod = new DimensionEffectsModifier(Optional.empty(), Optional.empty(), Optional.empty(), Optional.empty(),
-                        Optional.empty(), Optional.ofNullable(fogCol), Optional.ofNullable(skyCol),
+                        Optional.empty(), Optional.ofNullable(fogCol), Optional.ofNullable(skyCol), Optional.empty(),
                         false, false, Optional.empty(), Set.of());
 
                 converted.put(ResourceLocation.parse(names[i]), mod);